+++ src/Pim/Bundle/FilterBundle/Filter/Flexible/EntityFilter.php
@@ -6,6 +6,7 @@ use Doctrine\Common\Collections\ArrayCollection;
 use Oro\Bundle\FilterBundle\Filter\EntityFilter as OroEntityFilter;
 use Oro\Bundle\FilterBundle\Datasource\FilterDatasourceAdapterInterface;
 use Pim\Bundle\FlexibleEntityBundle\Manager\FlexibleManager;
+use Pim\Bundle\FilterBundle\Filter\ProductFilterUtility;
 
 /**
  * Flexible filter
@@ -41,8 +42,8 @@ class EntityFilter extends OroEntityFilter
 
         $this->util->applyFlexibleFilter(
             $ds,
-            $this->get(FilterUtility::FEN_KEY),
-            $this->get(FilterUtility::DATA_NAME_KEY),
+            $this->get(ProductFilterUtility::FEN_KEY),
+            $this->get(ProductFilterUtility::DATA_NAME_KEY),
             $this->extractIds($data['value']),
             $operator
         );
@@ -60,7 +61,7 @@ class EntityFilter extends OroEntityFilter
     protected function getClassName()
     {
         /** @var FlexibleManager $fm */
-        $flexibleManager = $this->util->getFlexibleManager($this->get(FilterUtility::FEN_KEY));
+        $flexibleManager = $this->util->getFlexibleManager($this->get(ProductFilterUtility::FEN_KEY));
         $valueName       = $flexibleManager->getFlexibleValueName();
         $valueMetadata   = $flexibleManager->getStorageManager()
             ->getMetadataFactory()