+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 - Move src/Pim/Bundle/FilterBundle/Filter/ProductGroupsFilter.php -> src/Pim/Bundle/FilterBundle/Filter/Product/GroupsFilter.php
 - Move src/Pim/Bundle/FilterBundle/Resources/public/js/datafilter/filter/category-filter.js -> src/Pim/Bundle/FilterBundle/Resources/public/js/datafilter/filter/product_category-filter.js
 - Move src/Pim/Bundle/FilterBundle/Filter/CategoryFilter.php -> src/Pim/Bundle/FilterBundle/Filter/Product/CategoryFilter.php
+- Move src/Pim/Bundle/FilterBundle/Filter/Flexible/FilterUtility.php -> src/Pim/Bundle/FilterBundle/Filter/ProductFilterUtility.php
 
 # 1.0.0 - "Hare We Go" (2014-03-06)